Reorganized commands, erased individual cogs
