+++ .gitignore
@@ -11,3 +11,4 @@ databases/config.toml
 robotito.log
 runtime.txt
 Procfile
+main_bot.py
