+++ cog/functions/__init__.py
@@ -1,3 +1,3 @@
 from cog.functions.functions import (rbColor, gif, header, sentence,
                                      get_8ball, get_penis, get_love,
-                                     get_love_gif, get_api)
+                                     get_love_gif)
