+++ cog/functions/functions.py
@@ -1,10 +1,10 @@
 import random
 import json
-import tomli
 import discord
 
 
 def open_str():
+    """Opens the string's database."""
     with open('databases/db_str.json', encoding='utf8') as f:
         data = f.read()
         string = json.loads(data)
@@ -15,6 +15,7 @@ string = open_str()
 
 
 def rbColor():
+    """Returns one of the main colors of RoboTito."""
     colors = [discord.Color.from_rgb(255, 94, 43),
               discord.Color.from_rgb(82, 92, 253),
               discord.Color.from_rgb(72, 159, 181),
@@ -25,19 +26,22 @@ def rbColor():
 # Functions used in cog/anime.py
 
 
-def gif(string):
+def gif(table):
+    """Returns an random gif from the given table."""
     with open('databases/db_gifs.json') as f:
         data = f.read()
         gif = json.loads(data)
-        return random.choice(gif[string])
+        return random.choice(gif[table])
 
 
 def header(table):
+    """Returns the "header" (main sentence) for interaction commands."""
     header = open_str()[table]
     return random.choice(header)
 
 
 def sentence(author, msg, member=None):
+    """Returns the message that will show in interaction commands."""
     if member is None:
         message = random.choice(string[msg])
         result = author + message
@@ -52,10 +56,12 @@ def sentence(author, msg, member=None):
 
 
 def get_8ball():
+    """Returns a random response for 8ball command."""
     return random.choice(string['ball8'])
 
 
 def get_penis(number: int):
+    """Returns a random measure unit for penis commands."""
     if number >= 1:
         return random.choice(string['penis'])
     else:
@@ -63,6 +69,7 @@ def get_penis(number: int):
 
 
 def get_love(number: int):
+    """Returns a random phrase for love command, based on the rate."""
     if number >= 75:
         return random.choice(string['love_high'])
     elif number >= 45:
@@ -72,6 +79,7 @@ def get_love(number: int):
 
 
 def get_love_gif(number: int):
+    """Returns a random gif for love command, based on the rate."""
     with open('databases/db_gifs.json', encoding='utf-8') as f:
         gifData = f.read()
         gifstring = json.loads(gifData)
@@ -81,13 +89,3 @@ def get_love_gif(number: int):
         else:
             gif = random.choice(gifstring['love_low'])
             return gif
-
-
-# REST API's Token
-
-
-def get_api():
-    with open('databases/config.toml', encoding='utf-8') as f:
-        token_data = tomli.load(f)
-        token = token_data['api']
-        return token
